revert reorg of APIVersion schema
